+++ Properties/Settings.Designer.cs
@@ -264,18 +264,6 @@ namespace Wrok.Properties {
 
         [global::System.Configuration.UserScopedSettingAttribute()]
         [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
-        [global::System.Configuration.DefaultSettingValueAttribute("False")]
-        public bool SettingsUpgraded {
-            get {
-                return ((bool)(this["SettingsUpgraded"]));
-            }
-            set {
-                this["SettingsUpgraded"] = value;
-            }
-        }
-
-        [global::System.Configuration.UserScopedSettingAttribute()]
-        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
         [global::System.Configuration.DefaultSettingValueAttribute("True")]
         public bool CheckForUpdates {
             get {